+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Do not allow `:amazon`, `amazon_sqs`, or `amazon_sqs_async` for SQS active job configuration. Instead use `:sqs` and `:sqs_async`.
+
 3.13.0 (2024-06-06)
 ------------------
 
@@ -92,7 +94,7 @@ Unreleased Changes
 3.4.0 (2020-12-07)
 ------------------
 
-* Feature - Add a non-blocking async ActiveJob adapter: `:sqs_async`.
+* Feature - Add a non-blocking async ActiveJob adapter: `:amazon_sqs_async`.
 
 * Feature - Add a lambda handler for processing active jobs from an SQS trigger.